+++ src/views/IndexView.vue
@@ -6,7 +6,7 @@ import pushContentData from "@/page-contents/index"
 import { onBeforeMount, onMounted } from 'vue';
 import { contentData } from '@/common/contentData';
 import ContentInfoCard from '@/components/ContentInfoCard.vue';
-import ContentList from '@/components/ContentList.vue';
+import ContentList from '@/components/ContentsList.vue';
 
 onBeforeMount(() => {
 	contentData.splice(0);
